Fix builds, given the latest libnds

+++ source/text.cpp
@@ -888,7 +888,7 @@ int asciiart() {
   return error;
 }
 
-const char* ErrorString(uint c) {
+const char* ErrorString(u8 c) {
 	switch (c) {
 		case 0:
 		return "ok";
